Replace softfail of bsc1204176 with poo reference
Bug bsc#1204176 has been marked as WONTFIX by the yast developers.
The bug can be only reproduced in the openQA environment and, as such,
it is now considered an openQA framework issue.
The resulting progress issue ticke is poo#124652 and it will be the
new reference point for the gtk refresh issue workaround.