Move the container to the devel project
The container needed for this was in Ricardo Branco's home OBS project and would
eventually be garbage collected.

This uses the same container that has been submitted to the CI subproject in the
devel namespace and is for this reason not going to be garbage collected.

Related ticket: https://jira.suse.com/browse/TEAM-7428