Use better variables for MPI module to avoid interchange values
When `openmpi` is used, because the various versions on each product we try to
get the required one using `get_mpi` function, despite the MPI job variable.
The mpi_master uses it but some operation need plain `openmpi`. Before the
implementation reused the $mpi variable but that had to interchange for one
step to the other. Using a new variable `mpi2load` it stores the value to be
used for loading the mpi module and let untouched the $mpi which can be used
in other steps later on.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>