lib/x11utils: ensure_unlocked_desktop: Don't wait until xscreensaver disappears
Previously it pressed a key to "update the screen" and then did a
wait_still_screen which timed out because of the animated xscreensaver prompt.
That caused the prompt to disappear again before it entered the password.

Ideally the second wait_still_screen gets removed, but to avoid introducing
some regression somewhere the timeout is halved instead.