Add smoke test for libica
The icainfo output also indicates, whether the libica library has
built-in FIPS support, whether it is running in FIPS mode, and whether
it is in an error state. Algorithms that are not FIPS approved are
marked as blocked in both table columns and cannot be processed when
running in FIPS mode. All algorithms are marked as blocked when libica
is in an error state.

- ticket: [Create new test to validate libica package in SLE Micro-s390x on FIPS](https://progress.opensuse.org/issues/125780)