Use ovmf xen firmware for uefi guest
Recent uefi guest installation on 15-SP5 xen host failed because
text console can not be opened for yast2 installation, which has
something to do with ovmf firmware used. After looking into ovmf
firmware definition, it seems that ovmf-x86_64-xen-4m.bin is the
right one to be used because it is used for xenfv virtual machine
specifically.