Replace NFS server 1c119 FQHN with its IP address
As part of ongoing works in our internal network, the FQHN
1c119.qa.suse.de has been decommissioned so NFS access to this server is
currently only possible using its IP address. This commit removes the
setting from the sles4sap schedules as it makes more sense to have such
a setting in a Test Suite or in the Job Group configuration; setting is
to be added in the Job Groups YAML configuration for tests using:

- schedule/sles4sap/netweaver/netweaver_single_systemd.yaml
- schedule/sles4sap/netweaver/test_netweaver.yaml

While also replacing the IP address instead of the FQHN in other team's
schedules.